Enhancement: Piggyback on existing SecurityDetails interfaces.
