+++ src/SecurityCompare/Converters/AssetAllocationsConverter.ts
@@ -23,10 +23,10 @@
 
 
 import {
-    SecurityCompareConverterOptions,
+    SecurityDetailsConverterOptions,
     SecurityCompareMetadata
-} from '../SecurityCompareOptions';
-import SecurityCompareJSON from '../SecurityCompareJSON';
+} from '../../SecurityDetails/SecurityDetailsOptions';
+import SecurityDetailsJSON from '../../SecurityDetails/SecurityDetailsJSON';
 import SecurityCompareConverter from '../SecurityCompareConverter';
 
 /* *
@@ -47,7 +47,7 @@ export class AssetAllocationsConverter extends SecurityCompareConverter {
 
 
     public constructor (
-        options?: SecurityCompareConverterOptions
+        options?: SecurityDetailsConverterOptions
     ) {
         super(options);
 
@@ -75,7 +75,7 @@ export class AssetAllocationsConverter extends SecurityCompareConverter {
 
 
     public override parse (
-        options: SecurityCompareConverterOptions
+        options: SecurityDetailsConverterOptions
     ): void {
         const metadata = this.metadata,
             ids = [],
@@ -88,7 +88,7 @@ export class AssetAllocationsConverter extends SecurityCompareConverter {
             json = userOptions.json;
 
         // Validate JSON
-        if (!SecurityCompareJSON.isSecurityCompareResponse(json)) {
+        if (!SecurityDetailsJSON.isSecurityCompareResponse(json)) {
             throw new Error('Invalid data');
         }
